+++ csrc/include/natten/cuda/fna_blackwell/kernel/sm100_fna_bwd_kernel_tma_warpspecialized.hpp
@@ -1880,7 +1880,9 @@ struct Sm100FnaBwdKernelTmaWarpSpecialized {
   CUTLASS_DEVICE void operator()(Params const& params, char* smem) {
 #if (                                            \
     !defined(CUTLASS_ARCH_MMA_SM100A_ENABLED) && \
-    !defined(CUTLASS_ARCH_MMA_SM100F_ENABLED))
+    !defined(CUTLASS_ARCH_MMA_SM100F_ENABLED) && \
+    !defined(CUTLASS_ARCH_MMA_SM103A_ENABLED) && \
+    !defined(CUTLASS_ARCH_MMA_SM103F_ENABLED))
     printf(
         "ERROR : Arch conditional MMA instruction used without targeting appropriate compute capability. Aborting.\n");
 #else