+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 ## [Main branch]
 * Extended Attention (FMHA) functionality:
   * Causal mask, variable length: for now only supported in CUTLASS FMHA and Blackwell FMHA.
-  * GQA/MQA: for now only in Blackwell FMHA.
 * Torch.compile support added
     * All libnatten ops are now registered as torch ops, enabling full-graph compilation with NATTEN
         ops.
+* TokPerm kernels: Moved dilation to batch instead of heads, which finally unblocks GQA/MQA.
+* GQA/MQA available in CUTLASS FNA/FMHA, Blackwell FNA/FMHA, and Flex FNA/FMHA.
 
 ## [0.21.1] - 2025-10-26
 * FP8 inference support for Blackwell DC-class ([#266](https://github.com/SHI-Labs/NATTEN/pull/266))